New architecture.
Classes MainWindow and glDisplay only handle the related ui widget.  Data is stored in a DTM or FlowPath object (both derives from Drawable base class).

glDisplay stores a reference to communicate with the main window, and const references to the data (that needs a reinterpret_cast between QList<FlowPath*> and QList<const FlowPath*>.
init() and draw() methods have been moved to private:, but a slot beginDraw has been written in replacement.
The drawing process begins with init(). Then, draw() calls drawData() on each object to be drawn. drawData calls the right (i.e. buildLegacy or buildArrays depending on the graphics primitive) drawing method. If arrays need to be built, as the object is constant, it cannot be achieved in glDisplay; thus a signal is emitted.
When a point is clicked in the viewer, DTM::ComputeIndex interpolates the coordinates on the DTM to find an index; this index is sent to MainWindow::addFlow that builds a new flow and adds it to the list. Succession of DTM points that makes the flow is computed inside the FlowPath class.
It seems that the inheritance of QGLViewer un glDisplay causes the link error.

MainWindow handles the operations on the data and other various UI processes.
The DTM is a pointer on an object, because it can change during the MainWindow session.
openDialog() empties the DTM (along with all flow paths, because they depends on the DTM used) and opens an other one.
rebuildArrays() calls the right method to build each data object's array.
deleteFlows is a small function called when all flows need to be destroyed.
Flows are iterated with an iterator most time.

draw.h defines two enumeration: the drawing mode, and the drawing primitive.
There is an helper function to get a string for debug messages, and an exception that should occur only when bad programming.
The helper function glPrimitive translates the enum primitive into the right OpenGL primitive.

Both DTM and FlowPath derives from Drawable class.
The Drawable class stores the vertices (while letting derivated classes handle it), and provides the interface to build and get the vertex array and the indices array.
Derivated classes cannot directly modify the vertex array and the indices array; they are stored as basic arrays because their size can be computed berforehand.
The Drawable class provides a build function (and a const overload) that can be called by child classes at building time. This build function uses the right mode.
The two top-level building functions are buildArrays() and buildLegacy() const. They are pure virtuals because only derivated classes have enough information to handle the global build process.
Vertex arrays and indices arrays have to be rebuilt at each mode or primitive change.
Copy constructor makes a deep copy that could fail if allocation fails. In that case, a std::bad_alloc is thrown.
USER: to customize the building once you have all your m_vertices, you should first call prepareBuild() that will reallocate arrays. Then, iterate your nodes in the order you want and simply call buildFunction.

DTM uses the parent's m_vertices. It has some mode infos and their getters.
Methods computeIndex is called by MainWindow when a point is entered or clicked.
readDTM, computeLineLength and computeDataSize are well known.
Other functions handles the building / drawing process. The super buildArrays() and buildLegacy() const are redefined. Template functions contains code specific for each primitive. And array_size computes the size of arrays for the current primitive.
The template functions are defined in a .inl file that do not need to be included when using the class declaration but has to be included when explicitely calling one of the functions (a priori, this only happens in dtm.cpp). A macro allows to define both const and non-const templates in a single line. For each primitive, an ARRAY_SIZE macro is defined (then used in the array_size function); its definition can call NB_CALL_LINE and NB_CALL_BACK macros that are defined in dtm.cpp and represents the number of times each template function is called.
In dtm.cpp, implementation of buildArrays and buildLegacy use the same code: it has been stored in macros to be written once.

FlowPath is simpler: the path is computed at the object construction and is not modified afterwards.
Indices are stored in m_minIndices and the corresponding vertices in m_vertices. The vertex array is thus m_vertices, indexed by [0..n].